Further simplify diagnostics crate
This is part of the simplification process we started a while ago. The
core principle is that datamodel diagnostics are not meant to be handled
in Prisma code, they are a regular output of the validation process and
should be optimized for ease of creation and displaying.
