--- libs/datamodel/diagnostics/src/validated.rs
@@ -1,7 +1,0 @@
-use crate::DatamodelWarning;
-
-#[derive(Debug, PartialEq, Clone)]
-pub struct Validated<T> {
-    pub subject: T,
-    pub warnings: Vec<DatamodelWarning>,
-}
